@@ -1,16 +1,22 @@
 apply plugin: 'com.android.application'
 
 android {
-    compileSdkVersion 28
-    // buildToolsVersion "26.0.0"
+    namespace 'com.vone.qrcode'
+    compileSdk 34
+    
     defaultConfig {
         applicationId "com.vone.qrcode"
-        minSdkVersion 19
-        targetSdkVersion 28
-        versionCode 2
-        versionName "2.1"
-        testInstrumentationRunner "android.support.test.runner.AndroidJUnitRunner"
+        minSdk 23
+        targetSdk 34
+        versionCode 3
+        versionName "3.0"
+        testInstrumentationRunner "androidx.test.runner.AndroidJUnitRunner"
     }
+    
+    buildFeatures {
+        buildConfig true
+    }
+    
     signingConfigs {
         //加载资源
         Properties properties = new Properties()
@@ -27,27 +33,44 @@ android {
     buildTypes {
         debug {
             signingConfig signingConfigs.signConfig // 配置debug包的签名,接入微信分享必须验证签名
+            buildConfigField "boolean", "ENABLE_SPLASH_TEST_TOOLS", "true"
+            manifestPlaceholders = [enableSplashTestTools: "true"]
         }
         release {
             minifyEnabled false
             proguardFiles getDefaultProguardFile('proguard-android.txt'), 'proguard-rules.pro'
             signingConfig signingConfigs.signConfig
+            buildConfigField "boolean", "ENABLE_SPLASH_TEST_TOOLS", "false"
+            manifestPlaceholders = [enableSplashTestTools: "false"]
         }
     }
-    lintOptions {
+    
+    compileOptions {
+        sourceCompatibility JavaVersion.VERSION_1_8
+        targetCompatibility JavaVersion.VERSION_1_8
+    }
+    
+    lint {
         abortOnError false
     }
 }
 
 dependencies {
     implementation fileTree(dir: 'libs', include: ['*.jar'])
-    androidTestImplementation('com.android.support.test.espresso:espresso-core:2.2.2', {
-        exclude group: 'com.android.support', module: 'support-annotations'
-    })
-    implementation 'com.android.support:appcompat-v7:28.0.0'
-    // implementation 'com.android.support.constraint:constraint-layout:1.0.2'
-    implementation 'com.google.zxing:core:3.3.0'
-    implementation 'com.squareup.okhttp3:okhttp:3.7.0'
-    implementation 'com.squareup.okio:okio:1.12.0'
-    testImplementation 'junit:junit:4.12'
+    androidTestImplementation 'androidx.test.espresso:espresso-core:3.5.1'
+    androidTestImplementation 'androidx.test.ext:junit:1.1.5'
+    androidTestImplementation 'androidx.test:runner:1.5.2'
+    androidTestImplementation 'androidx.test:rules:1.5.0'
+    androidTestImplementation 'androidx.test.espresso:espresso-contrib:3.5.1'
+    androidTestImplementation 'androidx.test.espresso:espresso-intents:3.5.1'
+    implementation 'androidx.appcompat:appcompat:1.6.1'
+    implementation 'androidx.constraintlayout:constraintlayout:2.1.4'
+    implementation 'com.google.zxing:core:3.5.2'
+    implementation 'com.squareup.okhttp3:okhttp:4.12.0'
+    implementation 'com.squareup.okio:okio:3.6.0'
+    implementation 'com.google.android.material:material:1.11.0'
+    implementation 'androidx.core:core-splashscreen:1.0.1'
+    testImplementation 'junit:junit:4.13.2'
+    testImplementation 'org.mockito:mockito-core:5.7.0'
+    testImplementation 'org.robolectric:robolectric:4.11.1'
 }
