@@ -66,20 +66,17 @@ public final class CaptureActivityHandler extends Handler {
 
   @Override
   public void handleMessage(Message message) {
-    switch (message.what) {
-      case R.id.auto_focus:
+    if (message.what == R.id.auto_focus) {
         //Log.d(TAG, "Got auto-focus message");
         // When one auto focus pass finishes, start another. This is the closest thing to
         // continuous AF. It does seem to hunt a bit, but I'm not sure what else to do.
         if (state == State.PREVIEW) {
           CameraManager.get().requestAutoFocus(this, R.id.auto_focus);
         }
-        break;
-      case R.id.restart_preview:
+    } else if (message.what == R.id.restart_preview) {
         Log.d(TAG, "Got restart preview message");
         restartPreviewAndDecode();
-        break;
-      case R.id.decode_succeeded:
+    } else if (message.what == R.id.decode_succeeded) {
         Log.d(TAG, "Got decode succeeded message");
         state = State.SUCCESS;
         Bundle bundle = message.getData();
@@ -90,24 +87,20 @@ public final class CaptureActivityHandler extends Handler {
         
         activity.handleDecode((Result) message.obj, barcode);//���ؽ��
         /***********************************************************************/
-        break;
-      case R.id.decode_failed:
+    } else if (message.what == R.id.decode_failed) {
         // We're decoding as fast as possible, so when one decode fails, start another.
         state = State.PREVIEW;
         CameraManager.get().requestPreviewFrame(decodeThread.getHandler(), R.id.decode);
-        break;
-      case R.id.return_scan_result:
+    } else if (message.what == R.id.return_scan_result) {
         Log.d(TAG, "Got return scan result message");
         activity.setResult(Activity.RESULT_OK, (Intent) message.obj);
         activity.finish();
-        break;
-      case R.id.launch_product_query:
+    } else if (message.what == R.id.launch_product_query) {
         Log.d(TAG, "Got product query message");
         String url = (String) message.obj;
         Intent intent = new Intent(Intent.ACTION_VIEW, Uri.parse(url));
         intent.addFlags(Intent.FLAG_ACTIVITY_CLEAR_WHEN_TASK_RESET);
         activity.startActivity(intent);
-        break;
     }
   }
 
